+++ NEWS.md
@@ -1,3 +1,12 @@
+# dqshiny 0.0.3
+
+* Fix unit test for new V8 (PR #1).
+* Support dqhot dataset changes and reconsider proper filters.
+* Change default filter for factors.
+* Support lists for filters specifying type and value.
+* Fix horizontal scroll only.
+* Fix initial column width alignment.
+
 # dqshiny 0.0.2
 
 * Fix text filter to properly skip empty filters.